feat(🔫🔌): enforce required plugin answers and harden runtime behavior
WHY: To keep team commit messages consistent and predictable.

SCOPES: Goodcommit / Builtin Plugin

Adds config-level required_answers overrides and enforces missing values as hard failures. splits the cli entrypoint into focused files and improves plugin lock portability, determinism, and verification. adds tests and docs for required answers, path resolution, and message parsing.
